# backend-safe-route/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Safe Route API
  version: 0.1.0
  description: Safe Route HTTP API. Business endpoints use the `/v1` prefix; `/health` reports service status.
servers:
  - url: http://127.0.0.1:8000
    description: Lokalny serwer developerski
...
paths:
...
  /v1/auth/login:
    post:
      tags: [Uwierzytelnianie]
      operationId: login
      summary: Zaloguj użytkownika
      description: Returns a short-lived JWT access token for the supplied email and password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              email: user@example.com
              password: correct-horse-battery-staple
      responses:
        "200":
          description: Użytkownik zalogowany.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
              example:
                access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                token_type: bearer
                expires_in: 900
        "401":
          description: Nieprawidłowy adres e-mail lub hasło.
          headers:
            WWW-Authenticate:
              $ref: "#/components/headers/WWWAuthenticate"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                detail: Invalid email or password
        "422":
          $ref: "#/components/responses/ValidationError"
        "503":
          description: Uwierzytelnianie nie jest skonfigurowane.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                detail: Authentication is not configured
  /v1/auth/register:
    post:
      tags: [Uwierzytelnianie]
      operationId: register
      summary: Utwórz konto użytkownika
      description: Tworzy konto lokalne i zwraca token oraz profil. Hasło ma 12–1024 znaki; nowe konto otrzymuje rolę `user`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
            example:
              email: ada@example.com
              password: a-long-unique-password
              given_name: Ada
              family_name: Lovelace
      responses:
        "201":
          description: Konto utworzone i użytkownik zalogowany.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationResponse"
              example:
                access_token: eyJhbGciOiJIUzI1NiJ9.example.signature
                token_type: bearer
                expires_in: 900
                user:
                  id: 42
                  email: ada@example.com
                  given_name: Ada
                  family_name: Lovelace
        "409":
          description: Konto z tym adresem e-mail już istnieje.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
        "503":
          $ref: "#/components/responses/AuthNotConfigured"
  /v1/auth/oauth/{provider}:
    post:
      tags: [Uwierzytelnianie]
      operationId: loginWithOAuth
      summary: Zaloguj przez Google lub Facebook
      description: Przyjmuje token ID Google lub token dostępu Facebooka w `credential`. Nowe konto jest tworzone automatycznie; istniejące konto nie jest łączone.
      parameters:
        - name: provider
          in: path
          required: true
          description: Dostawca tożsamości.
          schema:
            type: string
            enum: [google, facebook]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OAuthLoginRequest"
            examples:
              google:
                summary: Token ID Google
                value:
                  credential: google-id-token-from-mobile-sdk
              facebook:
                summary: Token dostępu Facebook
                value:
                  credential: facebook-access-token-from-mobile-sdk
      responses:
        "200":
          description: Użytkownik zalogowany lub konto utworzone.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationResponse"
              example:
                access_token: eyJhbGciOiJIUzI1NiJ9.example.signature
                token_type: bearer
                expires_in: 900
                user:
                  id: 42
                  email: ada@example.com
                  given_name: Ada
                  family_name: Lovelace
        "401":
          description: Token dostawcy jest nieprawidłowy lub konto nieaktywne.
          headers:
            WWW-Authenticate:
              $ref: "#/components/headers/WWWAuthenticate"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Istnieje konto z tym adresem, zaloguj się dotychczasową metodą.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
        "503":
          description: Dostawca OAuth lub konfiguracja OAuth są niedostępne.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
    WWWAuthenticate:
      description: Wymagany typ uwierzytelniania.
      schema:
        type: string
      example: Bearer
  responses:
...
    AuthNotConfigured:
      description: Uwierzytelnianie nie jest skonfigurowane.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            detail: Authentication is not configured
...
    ValidationError:
      description: Żądanie nie przeszło walidacji.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HTTPValidationError"
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          title: Email
          example: user@example.com
        password:
          type: string
          minLength: 1
          maxLength: 1024
          format: password
          writeOnly: true
          example: correct-horse-battery-staple
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: ada@example.com
        password:
          type: string
          minLength: 12
          maxLength: 1024
          format: password
          writeOnly: true
          example: a-long-unique-password
        given_name:
          type: [string, "null"]
          maxLength: 100
        family_name:
          type: [string, "null"]
          maxLength: 100
    OAuthLoginRequest:
      type: object
      required: [credential]
      properties:
        credential:
          type: string
          minLength: 1
          maxLength: 8192
          writeOnly: true
    TokenResponse:
      type: object
      required: [access_token, token_type, expires_in]
      properties:
        access_token:
          type: string
          description: Short-lived JWT for authorizing API requests.
        token_type:
          type: string
          const: bearer
        expires_in:
          type: integer
          description: Czas ważności tokenu w sekundach.
          minimum: 1
          example: 900
    AuthenticatedUser:
      type: object
      required: [id, email, given_name, family_name]
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        given_name:
          type: [string, "null"]
        family_name:
          type: [string, "null"]
    RegistrationResponse:
      type: object
      required: [access_token, token_type, expires_in, user]
      properties:
        access_token:
          type: string
          description: Krótkotrwały JWT dostępowy.
        token_type:
          type: string
          const: bearer
        expires_in:
          type: integer
          minimum: 1
        user:
          $ref: "#/components/schemas/AuthenticatedUser"
...
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
    HTTPValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            $ref: "#/components/schemas/ValidationError"
    ValidationError:
      type: object
      required: [loc, msg, type]
      properties:
        loc:
          type: array
          items:
            anyOf:
              - type: string
              - type: integer
        msg:
          type: string
        type:
          type: string
        input: {}
        ctx:
          type: object
          additionalProperties: true